requirements.txt
numpy
pandas
polars-lts-cpu
openpyxl
scipy
statsmodels
matplotlib
seaborn
scikit-learn
rdkit
python-dotenv
psycopg
sqlalchemy
jupyterlab
ipywidgets
jupyter_ai
streamlit
pygwalker
